% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/businessReport.R
\name{businessReport}
\alias{businessReport}
\title{Business Report (PDF)}
\usage{
businessReport(template = "INWTlab", resetStyleFiles = FALSE, ...)
}
\arguments{
\item{template}{character; Pandoc template to use for rendering. Pass
\code{"INWTlab"} to use the default example template}

\item{resetStyleFiles}{logical; should the style files (logo, cover, defs.tex) be overwritten with the default files?}

\item{...}{further arguments passed to \code{\link[rmarkdown]{pdf_document}}}
}
\description{
This function can be used to render the INWTlab business report. The design
and layout is loosly oriented at MS Word defaults.
}
\details{
The function serves as wrapper to \code{\link[rmarkdown]{pdf_document}}
only steering the selection of the template.
}
