% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-INWTlab_report.R
\name{INWTlab_report}
\alias{INWTlab_report}
\title{Report (PDF)}
\usage{
INWTlab_report(template = "INWTlab", ...)
}
\arguments{
\item{template}{character; Pandoc template to use for rendering. Pass
\code{"INWTlab"} to use the default example template}

\item{...}{further arguments passed to \code{\link[rmarkdown]{pdf_document}}}
}
\description{
Report (PDF)
}
\details{
The function serves as wrapper to \code{\link[rmarkdown]{pdf_document}}
only steering the selection of the template.
}
